acpi: video: avoid use edge values for bqc quirk
On ASUS N56VZ, its _BCL returns a package as:
{100, 40, 100, 99, 98, 97, ..., 3, 2, 1, 0}

The brightness level we set and get are:
set level to 100, _BQC returns 100;
set level t0 0, _BQC returns 100;
set level to 5, _BQC returns 4;
set level to 97, _BQC returns 96;
set level to 9, _BQC returns 8;
...

So if we use the max brightness level to do the test, we would think
_BQC of this system works OK; if we use the min brightness level to do
the test, we would think _BQC of this system is using index. To solve
this problem, we choose a test level that is not the same as the current
brightness level, and also not the same as either the max or min
brightness level. Hopefully, such a test level should be good enough to
work around most broken _BQC implementation, if not all.

Buglink: https://bugzilla.kernel.org/show_bug.cgi?id=52951
Reported-and-tested-by: Micael Dias <kam1kaz3@gmail.com>
Signed-off-by: Aaron Lu <aaron.lu@intel.com>