libata: acpi: set ACPI state for empty port
Previously we only set ACPI state for ports that have device attached.
Since it is desirable to power off unused port during runtime, this
patch changes the logic of sata_acpi_set_state such that for all port,
ACPI state will be set accordingly as long as there is an ACPI handle
associate with it.

Signed-off-by: Aaron Lu <aaron.lu@intel.com>