mm: convert zone lock from spinlock to rwlock
This patch converts zone lock from spinlock to rwlock and always
take the lock in write mode so there is no functionality change.

This is a preparation for the range lock patch which takes the lock
in read mode to make allocation and free path work concurrently.

compact_trylock and compact_unlock_should_abort are taken from
Daniel Jordan's patch.

Per Dave's suggestion, I tested single thread will-it-scale/page_fault1
performance to see if rwlock's performance is worse on a desktop:

v4.19-no-lru
average: 338885

this patch(rwlock)
average: 338564

It's about the same.

Signed-off-by: Aaron Lu <aaron.lu@intel.com>